Updates to vector.
normalise no longer modifies the input
vectors in place. Use the 'out' parameter or
the return value to receive the result.
set_length no longer modifies the input
vectors in place. Use the 'out' parameter or
the return value to receive the result.
Simplify some function logic.
Use the length function instead of calculating
it for numerous functions.
Rename some parameters for consistency.